Rename domain classifiers to purpose-based names
classify_domains_claude.py -> classify_domains_primary.py
classify_batch_bedrock.py  -> classify_domains_other.py

Names now describe role (primary vs 'Other' bucket) rather than backend.
Update the import in fetch_readmes_for_classification.py and references in
README, .env.template, and pyproject.toml.
