+++ .env.template
@@ -7,7 +7,7 @@
 # GitHub API token: repo discovery + README fetch (avoids the 60/hr public rate limit).
 GITHUB_TOKEN=ghp_your_token_here
 
-# Anthropic API key: primary domain classifier (analysis/data_collection/classify_domains_claude.py).
+# Anthropic API key: primary domain classifier (analysis/data_collection/classify_domains_primary.py).
 ANTHROPIC_API_KEY=sk-ant-your_key_here
 
 # Optional: alternate LLM classifiers kept in analysis/archive/ (not needed to reproduce the paper).
